+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 Bugfixes:
   - Indexes do not work in objects named "properties" [#697](https://github.com/pubkey/rxdb/issues/697)
+  - Wrong pouch-location when folderpath used for collection [#677](https://github.com/pubkey/rxdb/issues/677)
 
 Other:
   - Updated Angular-Example to 6.0.5 Thanks [@fuerst](https://github.com/fuerst)